feat: Update API_BASE_URL based on environment variable
This commit updates the `API_BASE_URL` in the `common.js` file to be determined based on the `ENVIRONMENT` variable from the `@env` module. If the `ENVIRONMENT` is set to "production", the `API_BASE_URL` will be set to the value of `PROD_API`. Otherwise, it will be set to `${LOCALHOST}/api/`.